Move content_crypt to internal
